+++ Sources/SuperwallKit/StoreKit/Transactions/TransactionManager.swift
@@ -16,6 +16,7 @@ final class TransactionManager {
   private let purchaseController: PurchaseController
   private let sessionEventsManager: SessionEventsManager
   private let eventsQueue: EventsQueue
+  private let productsFetcher: ProductsFetcherSK1
   private let factory: Factory
   typealias Factory = OptionsFactory
     & TriggerFactory
@@ -23,6 +24,10 @@ final class TransactionManager {
     & StoreTransactionFactory
     & DeviceHelperFactory
     & HasExternalPurchaseControllerFactory
+  enum State {
+    case observing
+    case purchasing(PurchaseSource)
+  }
 
   init(
     storeKitManager: StoreKitManager,
@@ -30,6 +35,7 @@ final class TransactionManager {
     purchaseController: PurchaseController,
     sessionEventsManager: SessionEventsManager,
     eventsQueue: EventsQueue,
+    productsFetcher: ProductsFetcherSK1,
     factory: Factory
   ) {
     self.storeKitManager = storeKitManager
@@ -37,10 +43,13 @@ final class TransactionManager {
     self.purchaseController = purchaseController
     self.sessionEventsManager = sessionEventsManager
     self.eventsQueue = eventsQueue
+    self.productsFetcher = productsFetcher
     self.factory = factory
   }
 
   /// Purchases the given product and handles the result appropriately.
+  ///
+  /// This uses a `PurchasingCoordinator` to coordinate the purchasing from start to finish.
   ///
   /// - Parameters:
   ///   - productId: The ID of the product to purchase.
@@ -61,35 +70,48 @@ final class TransactionManager {
         return .failed(PurchaseError.productUnavailable)
       }
       product = storeProduct
-    case .external(let storeProduct):
+    case .purchaseFunc(let storeProduct):
       product = storeProduct
+    case .observeFunc:
+      // This is a no-op, there's no way someone can call this func using observe.
+      return .cancelled
     }
-    let isEligibleForFreeTrial = await receiptManager.isFreeTrialAvailable(for: product)
-
-    await prepareToPurchase(product: product, purchaseSource: purchaseSource)
+    await prepareToPurchase(
+      product: product,
+      purchaseSource: purchaseSource
+    )
 
     let result = await purchase(product, purchaseSource: purchaseSource)
 
     // Return early if using a purchase controller and purchasing externally.
     // This avoids duplicate calls by the purchase function of the purchase
     // controller.
-    if case .external = purchaseSource,
+    if case .purchaseFunc = purchaseSource,
       factory.makeHasExternalPurchaseController() {
+      // Not resetting coordinator here because we need it with the purchase controller
+      // call.
       return result
     }
 
+    await handle(
+      result: result,
+      state: .purchasing(purchaseSource)
+    )
+
+    return result
+  }
+
+  func handle(
+    result: PurchaseResult,
+    state: State
+  ) async {
+    let coordinator = factory.makePurchasingCoordinator()
+
     switch result {
     case .purchased:
-      await didPurchase(
-        product: product,
-        purchaseSource: purchaseSource,
-        didStartFreeTrial: isEligibleForFreeTrial
-      )
+      await didPurchase()
     case .restored:
-      await didRestore(
-        product: product,
-        restoreSource: purchaseSource.toRestoreSource()
-      )
+      await didRestore()
     case .failed(let error):
       let superwallOptions = factory.makeSuperwallOptions()
       guard let outcome = TransactionErrorLogic.handle(
@@ -97,41 +119,40 @@ final class TransactionManager {
         triggers: factory.makeTriggers(),
         shouldShowPurchaseFailureAlert: superwallOptions.paywalls.shouldShowPurchaseFailureAlert
       ) else {
-        await trackFailure(
-          error: error,
-          product: product,
-          purchaseSource: purchaseSource
-        )
-        if case let .internal(_, paywallViewController) = purchaseSource {
-          await paywallViewController.togglePaywallSpinner(isHidden: true)
+        await trackFailure(error: error)
+        switch state {
+        case .observing:
+          break
+        case .purchasing(let purchaseSource):
+          if case let .internal(_, paywallViewController) = purchaseSource {
+            await paywallViewController.togglePaywallSpinner(isHidden: true)
+          }
         }
-        return result
+        return await coordinator.reset()
       }
       switch outcome {
       case .cancelled:
-        await trackCancelled(
-          product: product,
-          purchaseSource: purchaseSource
-        )
+        await trackCancelled()
       case .presentAlert:
-        await trackFailure(
-          error: error,
-          product: product,
-          purchaseSource: purchaseSource
-        )
-        await presentAlert(
-          title: "An error occurred",
-          message: error.safeLocalizedDescription,
-          source: purchaseSource.toGenericSource()
-        )
+        await trackFailure(error: error)
+        switch state {
+        case .observing:
+          break
+        case .purchasing(let purchaseSource):
+          await presentAlert(
+            title: "An error occurred",
+            message: error.safeLocalizedDescription,
+            source: purchaseSource.toGenericSource()
+          )
+        }
       }
     case .pending:
-      await handlePendingTransaction(purchaseSource: purchaseSource)
+      await handlePendingTransaction()
     case .cancelled:
-      await trackCancelled(product: product, purchaseSource: purchaseSource)
+      await trackCancelled()
     }
 
-    return result
+    await coordinator.reset()
   }
 
   @MainActor
@@ -255,17 +276,21 @@ final class TransactionManager {
     }
   }
 
-  private func didRestore(
-    product: StoreProduct? = nil,
-    restoreSource: RestoreSource
-  ) async {
-    let purchasingCoordinator = factory.makePurchasingCoordinator()
+  func didRestore(restoreSource: RestoreSource? = nil) async {
+    let coordinator = factory.makePurchasingCoordinator()
     var transaction: StoreTransaction?
     let restoreType: RestoreType
 
+    let coordinatorSource = await coordinator.source
+    guard let source = restoreSource ?? coordinatorSource?.toRestoreSource() else {
+      return
+    }
+
+    let product = await coordinator.product
+
     if let product = product {
       // Product exists so much have been via a purchase of a specific product.
-      transaction = await purchasingCoordinator.getLatestTransaction(
+      transaction = await coordinator.getLatestTransaction(
         forProductId: product.productIdentifier,
         factory: factory
       )
@@ -275,16 +300,21 @@ final class TransactionManager {
       restoreType = .viaRestore
     }
 
-    switch restoreSource {
+    var isObserved = false
+    if case .observeFunc = coordinatorSource {
+      isObserved = true
+    }
+
+    switch source {
     case .internal(let paywallViewController):
       let paywallInfo = await paywallViewController.info
-
       let trackedEvent = InternalSuperwallEvent.Transaction(
         state: .restore(restoreType),
         paywallInfo: paywallInfo,
         product: product,
-        model: transaction,
+        transaction: transaction,
         source: .internal,
+        isObserved: isObserved,
         storeKitVersion: .storeKit1
       )
       await Superwall.shared.track(trackedEvent)
@@ -299,8 +329,9 @@ final class TransactionManager {
         state: .restore(restoreType),
         paywallInfo: .empty(),
         product: product,
-        model: transaction,
+        transaction: transaction,
         source: .external,
+        isObserved: isObserved,
         storeKitVersion: .storeKit1
       )
       await Superwall.shared.track(trackedEvent)
@@ -314,15 +345,14 @@ final class TransactionManager {
     guard let sk1Product = product.sk1Product else {
       return .failed(PurchaseError.productUnavailable)
     }
-    await factory.makePurchasingCoordinator().beginPurchase(
-      of: product.productIdentifier,
-      source: purchaseSource
-    )
     switch purchaseSource {
     case .internal:
       return await purchaseController.purchase(product: sk1Product)
-    case .external:
+    case .purchaseFunc:
       return await factory.purchase(product: sk1Product)
+    case .observeFunc:
+      // No-op, there's no way this can be called from observe.
+      return .cancelled
     }
   }
 
@@ -332,12 +362,21 @@ final class TransactionManager {
 
   // MARK: - Transaction lifecycle
 
-  private func trackFailure(
-    error: Error,
-    product: StoreProduct,
-    purchaseSource: PurchaseSource
-  ) async {
-    switch purchaseSource {
+  func trackFailure(error: Error) async {
+    let coordinator = factory.makePurchasingCoordinator()
+    guard
+      let source = await coordinator.source,
+      let product = await coordinator.product
+    else {
+      return
+    }
+
+    var isObserved = false
+    if case .observeFunc = source {
+      isObserved = true
+    }
+
+    switch source {
     case .internal(_, let paywallViewController):
       Logger.debug(
         logLevel: .debug,
@@ -351,19 +390,21 @@ final class TransactionManager {
       )
 
       let paywallInfo = await paywallViewController.info
-      Task {
+      Task { [isObserved] in
         let trackedEvent = InternalSuperwallEvent.Transaction(
           state: .fail(.failure(error.safeLocalizedDescription, product)),
           paywallInfo: paywallInfo,
           product: product,
-          model: nil,
+          transaction: nil,
           source: .internal,
+          isObserved: isObserved,
           storeKitVersion: .storeKit1
         )
         await Superwall.shared.track(trackedEvent)
         await paywallViewController.webView.messageHandler.handle(.transactionFail)
       }
-    case .external:
+    case .purchaseFunc,
+      .observeFunc:
       Logger.debug(
         logLevel: .debug,
         scope: .paywallTransactions,
@@ -374,13 +415,14 @@ final class TransactionManager {
         error: error
       )
 
-      Task {
+      Task { [isObserved] in
         let trackedEvent = InternalSuperwallEvent.Transaction(
           state: .fail(.failure(error.safeLocalizedDescription, product)),
           paywallInfo: .empty(),
           product: product,
-          model: nil,
+          transaction: nil,
           source: .external,
+          isObserved: isObserved,
           storeKitVersion: .storeKit1
         )
         await Superwall.shared.track(trackedEvent)
@@ -388,11 +430,47 @@ final class TransactionManager {
     }
   }
 
+  func observeTransaction(for productId: String) async {
+    guard let storeProduct = try? await productsFetcher.products(
+      identifiers: [productId],
+      forPaywall: nil,
+      event: nil
+    ).first else {
+      Logger.debug(
+        logLevel: .debug,
+        scope: .superwallCore,
+        message: "There's a purchase happening of a product with id \(productId), "
+          + "but we couldn't retrieve it to observe its purchase."
+      )
+      return
+    }
+    let coordinator = factory.makePurchasingCoordinator()
+    await prepareToPurchase(
+      product: storeProduct,
+      purchaseSource: .observeFunc(storeProduct)
+    )
+    await coordinator.setCompletion { result in
+      Task { [weak self] in
+        await self?.handle(
+          result: result,
+          state: .observing
+        )
+      }
+    }
+  }
+
   /// Tracks the analytics and logs the start of the transaction.
-  private func prepareToPurchase(
+  func prepareToPurchase(
     product: StoreProduct,
     purchaseSource: PurchaseSource
   ) async {
+    let isFreeTrialAvailable = await receiptManager.isFreeTrialAvailable(for: product)
+
+    var isObserved = false
+    if case .observeFunc = purchaseSource {
+      isObserved = true
+    }
+
     switch purchaseSource {
     case .internal(_, let paywallViewController):
       Logger.debug(
@@ -409,8 +487,9 @@ final class TransactionManager {
         state: .start(product),
         paywallInfo: paywallInfo,
         product: product,
-        model: nil,
+        transaction: nil,
         source: .internal,
+        isObserved: isObserved,
         storeKitVersion: .storeKit1
       )
       await Superwall.shared.track(trackedEvent)
@@ -419,12 +498,16 @@ final class TransactionManager {
       await MainActor.run {
         paywallViewController.loadingState = .loadingPurchase
       }
-    case .external:
+    case .purchaseFunc,
+      .observeFunc:
       // If an external purchase controller is being used, skip because this will
       // get called by the purchase function of the purchase controller.
-      if factory.makeHasExternalPurchaseController() {
+      let options = factory.makeSuperwallOptions()
+      if !options.shouldObservePurchases,
+        factory.makeHasExternalPurchaseController() {
         return
       }
+
       Logger.debug(
         logLevel: .debug,
         scope: .paywallTransactions,
@@ -435,22 +518,36 @@ final class TransactionManager {
         state: .start(product),
         paywallInfo: .empty(),
         product: product,
-        model: nil,
+        transaction: nil,
         source: .external,
+        isObserved: isObserved,
         storeKitVersion: .storeKit1
       )
       await Superwall.shared.track(trackedEvent)
     }
+
+    await factory.makePurchasingCoordinator().beginPurchase(
+      of: product,
+      source: purchaseSource,
+      isFreeTrialAvailable: isFreeTrialAvailable
+    )
   }
 
   /// Dismisses the view controller, if the developer hasn't disabled the option.
-  private func didPurchase(
-    product: StoreProduct,
-    purchaseSource: PurchaseSource,
-    didStartFreeTrial: Bool
-  ) async {
-    switch purchaseSource {
+  func didPurchase() async {
+    let coordinator = factory.makePurchasingCoordinator()
+    guard
+      let source = await coordinator.source,
+      let product = await coordinator.product
+    else {
+      return
+    }
+
+    switch source {
     case .internal(_, let paywallViewController):
+      guard let product = await coordinator.product else {
+        return
+      }
       Logger.debug(
         logLevel: .debug,
         scope: .paywallTransactions,
@@ -469,13 +566,7 @@ final class TransactionManager {
       )
 
       await receiptManager.loadPurchasedProducts()
-
-      await trackTransactionDidSucceed(
-        transaction,
-        product: product,
-        purchaseSource: purchaseSource,
-        didStartFreeTrial: didStartFreeTrial
-      )
+      await trackTransactionDidSucceed(transaction)
 
       let superwallOptions = factory.makeSuperwallOptions()
       if superwallOptions.paywalls.automaticallyDismiss {
@@ -484,7 +575,8 @@ final class TransactionManager {
           result: .purchased(productId: product.productIdentifier)
         )
       }
-    case .external:
+    case .purchaseFunc,
+      .observeFunc:
       Logger.debug(
         logLevel: .debug,
         scope: .paywallTransactions,
@@ -503,21 +595,26 @@ final class TransactionManager {
 
       await receiptManager.loadPurchasedProducts()
 
-      await trackTransactionDidSucceed(
-        transaction,
-        product: product,
-        purchaseSource: purchaseSource,
-        didStartFreeTrial: didStartFreeTrial
-      )
+      await trackTransactionDidSucceed(transaction)
     }
   }
 
   /// Track the cancelled
-  private func trackCancelled(
-    product: StoreProduct,
-    purchaseSource: PurchaseSource
-  ) async {
-    switch purchaseSource {
+  func trackCancelled() async {
+    let coordinator = factory.makePurchasingCoordinator()
+    guard
+      let source = await coordinator.source,
+      let product = await coordinator.product
+    else {
+      return
+    }
+
+    var isObserved = false
+    if case .observeFunc = source {
+      isObserved = true
+    }
+
+    switch source {
     case .internal(_, let paywallViewController):
       Logger.debug(
         logLevel: .debug,
@@ -532,8 +629,9 @@ final class TransactionManager {
         state: .abandon(product),
         paywallInfo: paywallInfo,
         product: product,
-        model: nil,
+        transaction: nil,
         source: .internal,
+        isObserved: isObserved,
         storeKitVersion: .storeKit1
       )
       await Superwall.shared.track(trackedEvent)
@@ -542,7 +640,8 @@ final class TransactionManager {
       await MainActor.run {
         paywallViewController.loadingState = .ready
       }
-    case .external:
+    case .purchaseFunc,
+      .observeFunc:
       Logger.debug(
         logLevel: .debug,
         scope: .paywallTransactions,
@@ -555,16 +654,27 @@ final class TransactionManager {
         state: .abandon(product),
         paywallInfo: .empty(),
         product: product,
-        model: nil,
+        transaction: nil,
         source: .external,
+        isObserved: isObserved,
         storeKitVersion: .storeKit1
       )
       await Superwall.shared.track(trackedEvent)
     }
   }
 
-  private func handlePendingTransaction(purchaseSource: PurchaseSource) async {
-    switch purchaseSource {
+  func handlePendingTransaction() async {
+    let coordinator = factory.makePurchasingCoordinator()
+    guard let source = await coordinator.source else {
+      return
+    }
+
+    var isObserved = false
+    if case .observeFunc = source {
+      isObserved = true
+    }
+
+    switch source {
     case .internal(_, let paywallViewController):
       Logger.debug(
         logLevel: .debug,
@@ -580,13 +690,15 @@ final class TransactionManager {
         state: .fail(.pending("Needs parental approval")),
         paywallInfo: paywallInfo,
         product: nil,
-        model: nil,
+        transaction: nil,
         source: .internal,
+        isObserved: isObserved,
         storeKitVersion: .storeKit1
       )
       await Superwall.shared.track(trackedEvent)
       await paywallViewController.webView.messageHandler.handle(.transactionFail)
-    case .external:
+    case .purchaseFunc,
+      .observeFunc:
       Logger.debug(
         logLevel: .debug,
         scope: .paywallTransactions,
@@ -598,8 +710,9 @@ final class TransactionManager {
         state: .fail(.pending("Needs parental approval")),
         paywallInfo: .empty(),
         product: nil,
-        model: nil,
+        transaction: nil,
         source: .external,
+        isObserved: isObserved,
         storeKitVersion: .storeKit1
       )
       await Superwall.shared.track(trackedEvent)
@@ -608,7 +721,7 @@ final class TransactionManager {
     await presentAlert(
       title: "Waiting for Approval",
       message: "Thank you! This purchase is pending approval from your parent. Please try again once it is approved.",
-      source: purchaseSource.toGenericSource()
+      source: source.toGenericSource()
     )
   }
 
@@ -644,95 +757,85 @@ final class TransactionManager {
     }
   }
 
-  func trackTransactionDidSucceed(
-    _ transaction: StoreTransaction?,
-    product: StoreProduct,
-    purchaseSource: PurchaseSource,
-    didStartFreeTrial: Bool
-  ) async {
-    switch purchaseSource {
+  func trackTransactionDidSucceed(_ transaction: StoreTransaction?) async {
+    let coordinator = factory.makePurchasingCoordinator()
+    guard
+      let source = await coordinator.source,
+      let product = await coordinator.product
+    else {
+      return
+    }
+    let didStartFreeTrial = await coordinator.isFreeTrialAvailable
+
+    var isObserved = false
+    if case .observeFunc = source {
+      isObserved = true
+    }
+
+    let type: InternalSuperwallEvent.Transaction.TransactionType = {
+      if product.subscriptionPeriod == nil {
+        return .nonRecurringProductPurchase
+      } else if didStartFreeTrial {
+        return .freeTrialStart
+      } else {
+        return .subscriptionStart
+      }
+    }()
+
+    let paywallInfo: PaywallInfo
+    let eventSource: InternalSuperwallEvent.Transaction.Source
+    switch source {
     case .internal(_, let paywallViewController):
-      let paywallShowingFreeTrial = await paywallViewController.paywall.isFreeTrialAvailable == true
-      let didStartFreeTrial = product.hasFreeTrial && paywallShowingFreeTrial
-
-      let paywallInfo = await paywallViewController.info
-
-      let trackedEvent = InternalSuperwallEvent.Transaction(
-        state: .complete(product, transaction),
-        paywallInfo: paywallInfo,
-        product: product,
-        model: transaction,
-        source: .internal,
-        storeKitVersion: .storeKit1
-      )
-      await Superwall.shared.track(trackedEvent)
+      paywallInfo = await paywallViewController.info
+      eventSource = .internal
       await paywallViewController.webView.messageHandler.handle(.transactionComplete)
+    case .purchaseFunc,
+      .observeFunc:
+      paywallInfo = .empty()
+      eventSource = .external
+    }
 
-      // Immediately flush the events queue on transaction complete.
-      await eventsQueue.flushInternal()
+    let trackedTransactionEvent = InternalSuperwallEvent.Transaction(
+      state: .complete(product, transaction, type),
+      paywallInfo: paywallInfo,
+      product: product,
+      transaction: transaction,
+      source: eventSource,
+      isObserved: isObserved,
+      storeKitVersion: .storeKit1
+    )
+    await Superwall.shared.track(trackedTransactionEvent)
+    await eventsQueue.flushInternal()
 
-      if product.subscriptionPeriod == nil {
-        let trackedEvent = InternalSuperwallEvent.NonRecurringProductPurchase(
+    switch type {
+    case .nonRecurringProductPurchase:
+      await Superwall.shared.track(
+        InternalSuperwallEvent.NonRecurringProductPurchase(
           paywallInfo: paywallInfo,
-          product: product
+          product: product,
+          transaction: transaction
         )
-        await Superwall.shared.track(trackedEvent)
-      } else {
-        if didStartFreeTrial {
-          let trackedEvent = InternalSuperwallEvent.FreeTrialStart(
-            paywallInfo: paywallInfo,
-            product: product
-          )
-          await Superwall.shared.track(trackedEvent)
-
-          let notifications = paywallInfo.localNotifications.filter {
-            $0.type == .trialStarted
-          }
-
-          await NotificationScheduler.scheduleNotifications(notifications, factory: factory)
-        } else {
-          let trackedEvent = InternalSuperwallEvent.SubscriptionStart(
-            paywallInfo: paywallInfo,
-            product: product
-          )
-          await Superwall.shared.track(trackedEvent)
-        }
-      }
-    case .external:
-      let trackedEvent = InternalSuperwallEvent.Transaction(
-        state: .complete(product, transaction),
-        paywallInfo: .empty(),
-        product: product,
-        model: transaction,
-        source: .external,
-        storeKitVersion: .storeKit1
       )
-      await Superwall.shared.track(trackedEvent)
-
-      // Immediately flush the events queue on transaction complete.
-      await eventsQueue.flushInternal()
-
-      if product.subscriptionPeriod == nil {
-        let trackedEvent = InternalSuperwallEvent.NonRecurringProductPurchase(
-          paywallInfo: .empty(),
-          product: product
+    case .freeTrialStart:
+      await Superwall.shared.track(
+        InternalSuperwallEvent.FreeTrialStart(
+          paywallInfo: paywallInfo,
+          product: product,
+          transaction: transaction
         )
-        await Superwall.shared.track(trackedEvent)
-      } else {
-        if didStartFreeTrial {
-          let trackedEvent = InternalSuperwallEvent.FreeTrialStart(
-            paywallInfo: .empty(),
-            product: product
-          )
-          await Superwall.shared.track(trackedEvent)
-        } else {
-          let trackedEvent = InternalSuperwallEvent.SubscriptionStart(
-            paywallInfo: .empty(),
-            product: product
-          )
-          await Superwall.shared.track(trackedEvent)
-        }
+      )
+      let notifications = paywallInfo.localNotifications.filter {
+        $0.type == .trialStarted
       }
+      await NotificationScheduler.scheduleNotifications(notifications, factory: factory)
+    case .subscriptionStart:
+      await Superwall.shared.track(
+        InternalSuperwallEvent.SubscriptionStart(
+          paywallInfo: paywallInfo,
+          product: product,
+          transaction: transaction
+        )
+      )
     }
   }
 }