+++ Tests/SuperwallKitTests/Analytics/Internal Tracking/TrackingLogicTests.swift
@@ -298,6 +298,7 @@ final class TrackingLogicTests: XCTestCase {
 
     let messageHandler = PaywallMessageHandler(
       sessionEventsManager: dependencyContainer.sessionEventsManager,
+      receiptManager: dependencyContainer.receiptManager,
       factory: dependencyContainer
     )
     let webView = SWWebView(